refactor the code to make it more clear
