+++ hat/assets/js/apps/Iaso/domains/instances/compare/components/InstanceLogContentBasic.tsx
@@ -1,17 +1,11 @@
 import React, { FunctionComponent } from 'react';
-import {
-    Table,
-    TableBody,
-    TableCell,
-    TableRow,
-    TableHead,
-} from '@mui/material';
+import { Table, TableCell, TableRow, TableHead } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import { useSafeIntl, IntlFormatMessage } from 'bluesquare-components';
 
-import { formatLabel } from '../../utils';
 import { FileContent } from '../../types/instance';
 import MESSAGES from '../messages';
+import InstanceLogContentBodyTable from './InstanceLogContentBodyTable';
 
 type Props = {
     fileContent: FileContent;
@@ -27,21 +21,6 @@ const useStyles = makeStyles(theme => ({
         borderRight: 'none !important',
         // @ts-ignore
         borderBottom: `1px solid ${theme.palette.ligthGray.border}  !important`,
-    },
-    tableCell: {
-        backgroundColor: 'transparent',
-        borderTop: 'none !important',
-        borderLeft: 'none !important',
-        borderRight: 'none !important',
-        // @ts-ignore
-        borderBottom: `1px solid ${theme.palette.ligthGray.border}  !important`,
-    },
-    tableCellLabelName: {
-        // @ts-ignore
-        color: theme.palette.mediumGray.main,
-    },
-    highlightedRow: {
-        backgroundColor: '#f7eb9c',
     },
 }));
 
@@ -68,70 +47,10 @@ export const InstanceLogContentBasic: FunctionComponent<Props> = ({
                     </TableCell>
                 </TableRow>
             </TableHead>
-            <TableBody>
-                {fileContent.logA &&
-                    fileContent.logB &&
-                    fileContent?.fields.map(question => {
-                        const isRelevantQuestion = ![
-                            'meta',
-                            'instanceID',
-                        ].includes(question.name);
-                        const hasLogContent =
-                            fileContent.logA.json[question.name] ||
-                            fileContent.logB.json[question.name];
-                        const field = fileDescriptor?.children.find(
-                            child => child.name === question.name,
-                        );
-                        const isValuesDifferent =
-                            fileContent.logA.json[question.name] !==
-                            fileContent.logB.json[question.name];
-
-                        if (isRelevantQuestion && hasLogContent) {
-                            return (
-                                <TableRow
-                                    key={question.name}
-                                    className={
-                                        isValuesDifferent
-                                            ? classes.highlightedRow
-                                            : undefined
-                                    }
-                                >
-                                    <TableCell
-                                        className={classes.tableCell}
-                                        align="left"
-                                    >
-                                        <div>
-                                            {fileDescriptor?.children && field
-                                                ? formatLabel(field)
-                                                : question.name}
-                                        </div>
-                                        <span
-                                            className={
-                                                classes.tableCellLabelName
-                                            }
-                                        >
-                                            {question.name}
-                                        </span>
-                                    </TableCell>
-                                    <TableCell
-                                        className={classes.tableCell}
-                                        align="left"
-                                    >
-                                        {/* TO DO : display section titles for groups */}
-                                        {fileContent?.logA.json[question.name]}
-                                    </TableCell>
-                                    <TableCell
-                                        className={classes.tableCell}
-                                        align="left"
-                                    >
-                                        {fileContent?.logB.json[question.name]}
-                                    </TableCell>
-                                </TableRow>
-                            );
-                        }
-                        return null;
-                    })}
-            </TableBody>
+            <InstanceLogContentBodyTable
+                fileContent={fileContent}
+                fileDescriptor={fileDescriptor}
+            />
         </Table>
     );
 };
